more config changes, renames
- Changed file extension from .mts to .ts, because .mts is not necessary when package.json has type: module
- Modified tsconfig to use correct module and moduleResolution
- Updated README.md
