+++ README.md
@@ -3,5 +3,6 @@
 ## Set Up
 
 1. Open up a terminal (or cd into) the project's root folder
-2. Install all npm dependencies with `npm install --dev`
-3. Build and run the project using the `npm run start` command
+2. Install all npm dependencies with `npm install --include=dev`
+3. Build the project with the `npm run build` command
+4. Run the project with the `npm run start` command
