+++ src/assets/pages/index.html
@@ -5,7 +5,7 @@
 
         <link href = "/styles/style.css" rel = "stylesheet" type = "text/css"/>
 
-        <script defer src = "/scripts/local.mjs" type = "module"></script>
+        <script defer src = "/scripts/local.js" type = "module"></script>
     </head>
     <body>
