+++ src/shared/board/sq-coord.ts
