+++ src/shared/board/square.ts
@@ -1,5 +1,5 @@
 
-import { Piece_Ref } from "../piece/piece-ref.mjs";
+import { Piece_Ref } from "../piece/piece-ref.js";
 
 export class Square {
     public pieceRef?: Piece_Ref;
