+++ src/shared/piece/move.ts
@@ -1,5 +1,5 @@
 
-import { Sq_Coord } from "../board/sq-coord.mjs";
+import { Sq_Coord } from "../board/sq-coord.js";
 
 export interface Move {
     from: Sq_Coord,
